@@ -1,11 +1,6 @@
 name: ready_artisans_admin
 description: A new Flutter project.
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-
-# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
@@ -18,27 +13,31 @@ dependencies:
 
 
 
-  cupertino_icons: ^1.0.2
-  google_fonts: ^3.0.1
+  cupertino_icons: ^1.0.5
+  google_fonts: ^4.0.4
   url_strategy: ^0.2.0
-  firebase_core: ^2.4.0
-  firebase_auth: ^4.2.3
-  cloud_firestore: ^4.3.0
+  firebase_core: ^2.13.1
+  firebase_auth: ^4.6.2
+  cloud_firestore: ^4.8.0
   firebase_storage: ^11.0.9
   auto_route: ^5.0.4
-  font_awesome_flutter: ^10.3.0
-  flutter_smart_dialog: ^4.8.0
-  provider: ^6.0.5
+  font_awesome_flutter: ^10.4.0
+  flutter_smart_dialog: ^4.9.1
+  flutter_riverpod: ^2.3.6
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   vs_scrollbar: ^0.2.1
+  image_picker: ^0.8.7+1
+  responsive_table:
+    git:
+      url: https://github.com/Emmanuelfrimpong/flutter_responsive_table.git
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.1
   auto_route_generator: ^5.0.3
   build_runner: ^2.3.3
 
