@@ -12,17 +12,17 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.5
-  google_fonts: ^3.0.1
-  firebase_core: ^2.4.1
-  firebase_auth: ^4.2.4
-  cloud_firestore: ^4.3.1
-  firebase_storage: ^11.0.10
-  font_awesome_flutter: ^10.3.0
-  flutter_markdown: ^0.6.13
-  flutter_native_splash: ^2.2.16
-  flutter_smart_dialog: ^4.8.0
+  google_fonts: ^4.0.3
+  firebase_core: ^2.8.0
+  firebase_auth: ^4.3.0
+  cloud_firestore: ^4.4.5
+  firebase_storage: ^11.0.16
+  font_awesome_flutter: ^10.4.0
+  flutter_markdown: ^0.6.14
+  flutter_native_splash: ^2.2.19
+  flutter_smart_dialog: ^4.9.0+4
   provider: ^6.0.5
-  image_picker: ^0.8.6
+  image_picker: ^0.8.7
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   geocoding:
