@@ -1,22 +1,25 @@
+// ignore_for_file: file_names
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import '../utils/app_colors.dart';
 
-import '../app_colors.dart';
 
 class CustomButton extends StatefulWidget {
   const CustomButton(
       {Key? key,
-      required this.onPressed,
+      this.onPressed,
       required this.text,
-      this.color = secondaryColor,
-      this.padding = const EdgeInsets.symmetric(horizontal: 20, vertical: 12),
-      this.radius = 5})
+      this.color = primaryColor,
+      this.padding = const EdgeInsets.symmetric(horizontal: 15, vertical: 12),
+      this.radius = 5,
+      this.icon})
       : super(key: key);
-  final VoidCallback onPressed;
+  final VoidCallback? onPressed;
   final String text;
   final Color color;
   final EdgeInsets? padding;
   final double radius;
+  final IconData? icon;
 
   @override
   State<CustomButton> createState() => _CustomButtonState();
@@ -37,17 +40,26 @@ class _CustomButtonState extends State<CustomButton> {
       child: Container(
           padding: widget.padding,
           decoration: BoxDecoration(
-            color: onHover ? Colors.white : widget.color,
+            color: onHover ? Colors.transparent : widget.color,
             borderRadius: BorderRadius.circular(widget.radius),
-            border: Border.all(color: onHover ? widget.color : Colors.white),
+            border: Border.all(color: widget.color),
           ),
           child: Center(
-            child: Text(
-              widget.text,
-              style: GoogleFonts.nunito(
-                  color: onHover ? widget.color : Colors.white,
-                  fontSize: 16,
-                  fontWeight: FontWeight.w500),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                if (widget.icon != null)
+                  Icon(widget.icon!,
+                      size: 18, color: onHover ? widget.color : Colors.white),
+                if (widget.icon != null) const SizedBox(width: 10),
+                Text(
+                  widget.text,
+                  style: GoogleFonts.roboto(
+                      color: onHover ? widget.color : Colors.white,
+                      fontSize: 16,
+                      fontWeight: FontWeight.w700),
+                ),
+              ],
             ),
           )),
     );
