@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-
-import '../app_colors.dart';
-import 'custom_dropdown.dart';
+import '../utils/app_colors.dart';
 
 class CustomCheckBox extends StatelessWidget {
   const CustomCheckBox(
