@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 class ArtisansPage extends StatefulWidget {
   const ArtisansPage({super.key});
@@ -10,6 +11,42 @@ class ArtisansPage extends StatefulWidget {
 class _ArtisansPageState extends State<ArtisansPage> {
   @override
   Widget build(BuildContext context) {
-    return Container();
+    var size = MediaQuery.of(context).size;
+    return Container(
+      width: size.width,
+      padding: const EdgeInsets.all(10),
+      child: SingleChildScrollView(
+        child: Column(children: [
+          Container(
+            width: size.width,
+            height: size.width > 800 ? 100 : 30,
+            padding: const EdgeInsets.symmetric(horizontal: 25),
+            child: Row(children: [
+              if (size.width <= 800)
+                Text(
+                  'Home/Artisans',
+                  style: GoogleFonts.nunito(
+                      color: Colors.black45, fontWeight: FontWeight.bold),
+                ),
+              if (size.width > 800)
+                Image.asset(
+                  'assets/images/logo_h_t.png',
+                ),
+              const SizedBox(
+                width: 50,
+              ),
+              const Spacer(),
+              if (size.width > 800)
+                Text(
+                  'Home/Artisans',
+                  style: GoogleFonts.nunito(
+                      color: Colors.black45, fontWeight: FontWeight.bold),
+                )
+            ]),
+          ),
+          const SizedBox(height: 20),
+        ]),
+      ),
+    );
   }
 }
