@@ -1,40 +1,32 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+
+// ignore: constant_identifier_names
 enum Pages { Dashboard, Users, Artisans, Orders, Settings }
 
-class StateManagement extends ChangeNotifier {
-  double _sideWidth = 60;
-  double get sideWidth => _sideWidth;
-  void setSideWidth(double width) {
-    _sideWidth = width;
-    notifyListeners();
-  }
+final sidebarWithProvider = StateProvider<double>((ref) => 60);
+final sideBarOpenProvider = StateProvider<bool>((ref) => false);
+final scaffoldKeyProvider =
+    Provider<GlobalKey<ScaffoldState>>((ref) => GlobalKey<ScaffoldState>());
+final routerKeyProvider =
+    Provider<GlobalKey<AutoRouterState>>((ref) => GlobalKey<AutoRouterState>());
+final homePageChangeNotifierProvider =
+    StateNotifierProvider<HomePageChangeNotifierProvider, Pages>(
+        (ref) => HomePageChangeNotifierProvider());
 
-  bool isSideBarOpen = false;
-  bool get getSideBarOpen => isSideBarOpen;
-  void toggleSideBar() {
-    isSideBarOpen = !isSideBarOpen;
-    _sideWidth = 200;
-    notifyListeners();
-  }
+class HomePageChangeNotifierProvider extends StateNotifier<Pages> {
+  HomePageChangeNotifierProvider() : super(Pages.Dashboard);
 
-  var scaffoldKey = GlobalKey<ScaffoldState>();
-  GlobalKey<ScaffoldState> get getScaffoldKey => scaffoldKey;
-  final routerKey = GlobalKey<AutoRouterState>();
-  GlobalKey<AutoRouterState> get getRouterKey => routerKey;
-  Pages currentPage = Pages.Dashboard;
-  Pages get getSelectedPage => currentPage;
-
-  void navigateTo(PageRouteInfo page, Pages pageItem, {bool? falsePop}) {
-    final router = getRouterKey.currentState?.controller;
-    if (currentPage != pageItem) {
+  void changePage(PageRouteInfo page, Pages pageItem,  GlobalKey<AutoRouterState> routerKey,{bool? falsePop}) {
+    final router = routerKey.currentState?.controller;
+    if (state != pageItem) {
       router?.push(page);
-      currentPage = pageItem;
+      state = pageItem;
     } else if (falsePop == true) {
       router?.push(page);
-      currentPage = pageItem;
+      state = pageItem;
     }
-    notifyListeners();
   }
 }
