@@ -2,7 +2,7 @@ import 'dart:io';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
-import 'package:ready_artisans/Models/Users/Users.dart';
+import 'package:ready_artisans/Models/Users/users.dart';
 
 import '../Components/smart_dialog.dart';
 
