@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 import 'package:provider/provider.dart';
 import 'package:ready_artisans/Pages/Home/Components/my_jobs.dart';
-import 'package:ready_artisans/Providers/NavigationProvider.dart';
+import 'package:ready_artisans/Providers/navigation_provider.dart';
 import 'package:ready_artisans/generated/assets.dart';
-import '../../Components/CustomDrawer.dart';
+import '../../Components/custom_drawer.dart';
 import '../../Styles/app_colors.dart';
 import 'Components/home.dart';
 import 'Components/my_orders.dart';
