@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
-import 'package:ready_artisans/Providers/NavigationProvider.dart';
+import 'package:ready_artisans/Providers/navigation_provider.dart';
 
 import '../../Styles/app_colors.dart';
 import '../../generated/assets.dart';
