@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:ready_artisans/Providers/NavigationProvider.dart';
+import 'package:ready_artisans/Providers/navigation_provider.dart';
 import 'package:ready_artisans/Styles/app_colors.dart';
 
 import '../../generated/assets.dart';
