@@ -4,8 +4,8 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import '../../Components/custom_dropdown.dart';
 import '../../Components/text_inputs.dart';
-import '../../Models/Category.dart';
-import '../../Providers/NavigationProvider.dart';
+import '../../Models/category.dart';
+import '../../Providers/navigation_provider.dart';
 import '../../Styles/app_colors.dart';
 
 class ServiceDetails extends StatefulWidget {
